Preserve Last Modified Date on Onedrive
Onedrive provide a "fileSystemInfo resource type" [1] that contains properties that are reported by the device's local
file system for the local version of an item. To fetch the original last modified date we have to pass an additional
argument which will be stored server side and retrieved in a second moment.
In the uploadFile() method a new PATCH request have been added to update those info.
In the chunkedUploadFile() is sufficient add a property to the request.

When a CloudNode is created by the OnedriveCloudNodeFactory the fileSystemInfo are used.

[1] https://learn.microsoft.com/en-us/graph/api/resources/filesysteminfo?view=graph-rest-1.0#notes