Added handling for database corruption
If database corruption is detected the database is deleted and then recreated. Properly recreating and opening the new database involves invalidating any references to the old database and then running any initialization logic.

Added "Invalidatable" as wrapper for database references
Added "DatabaseOpenHelperFactory" to the dagger graph and added invalidation logic to "PatchedCallback"
Added dynamically delegating implementations for all entities to allow seamless swapping of underlying database instances
Configured Room to initialize based on the database template if a database with version 0 is encountered