Added tests for verifying BOMs
"PatchedCallback" and "DatabaseOpenHelperFactoryKt.patchConfiguration" interact with the inner classes of "SupportSQLiteOpenHelper" in a way that is prone to introducing bugs due to unnoticed changes to those classes.
The tests introduced by this commit ensure that changes to the bills of materials for those classes are noticed and handled in the future.