Update dependencies and migrate to Jetpack view binding
Kotlin upgrade was required due to some dependencies that uses OkHttp. This Kotlin upgrade required Gradle upgrade which required migration to Jetpack view binding.
As we touched a lot of the UI code due to the migration, we now use as ID always snake case and not camel case and snake case mixed.