UI hardening
Prioritize the visible area for downloading the files and generating the thumbnails. Add MAX_CONCURRENT_THUMBNAILS_THREADS to limit the number of usecase-threads generated. Do not start a double download of the same file and, finally, remove the progress bars for dowloads and substitute them with simple icons to lighten th UI.